Update README.md
My project is a color identifier. On Jupiter Notebook, once your camera is fully functional, you can choose a random object from your room and hold it up to the camera. Out of the rainbow colors, it will tell you which it is. This project is used to help the blind.